requirements.txt
django-imagekit
django-model-utils>=2.0
# Additional requirements go here
